refactor: extract oversized modal and page components (ADS-285)
ChatDetailModal (747→~130 lines): extract MessagesTab, ParticipantsTab,
DetailsTab, ModerationTab into ChatDetailModal/ folder. Tab-specific state
and hooks now live in the tab that owns them.

RescueDetailModal (843→~120 lines): extract OverviewTab, ContactTab,
PoliciesTab, StaffTab, ListingsTab, PlanTab into RescueDetailModal/ folder.
StaffTab owns its own fetching/mutations; PlanTab owns plan state.

SearchPage (563→~210 lines): extract useNearbySearch (wraps useGeolocation
+ geocoding), useSearchFilters (filter state + URL sync), and SearchFilters
component (filter card UI).

Add tests for MessagesTab, ModerationTab, StaffTab, PlanTab, useNearbySearch,
useSearchFilters, and SearchFilters. All 321 admin + 284 client tests pass.
