+++ app.admin/src/components/modals/RescueDetailModal/ListingsTab.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import * as styles from '../RescueDetailModal.css';
+
+export const ListingsTab: React.FC = () => (
+  <div className={styles.section}>
+    <h3 className={styles.sectionTitle}>Pet Listings</h3>
+    <div className={styles.infoValue}>Listings management coming soon...</div>
+  </div>
+);
