+++ app.admin/src/components/modals/index.ts
@@ -1,5 +1,5 @@
 // Rescue modals
-export { RescueDetailModal } from './RescueDetailModal';
+export { RescueDetailModal } from './RescueDetailModal/index';
 export { RescueVerificationModal } from './RescueVerificationModal';
 export { SendEmailModal } from './SendEmailModal';
 
@@ -10,7 +10,7 @@ export { CreateSupportTicketModal } from './CreateSupportTicketModal';
 export { UserActionsMenu } from './UserActionsMenu';
 
 // Chat modals
-export { ChatDetailModal } from './ChatDetailModal';
+export { ChatDetailModal } from './ChatDetailModal/index';
 
 // Bulk operations
 export { BulkConfirmationModal } from './BulkConfirmationModal';
